refactor: Migrate remaining common module tests to separate files
Extract inline test modules from string_similarity.rs and mod.rs into
dedicated test files under src/common/tests/. Maintains all 30 tests
(13 string_similarity + 17 common_mod) with identical behavior.
