+++ src/common/tests/mod.rs
@@ -1,1 +1,3 @@
 mod method_registry;
+mod string_similarity;
+mod common_mod;
